requirements.txt
appdirs==1.4.3
Django==1.10.7
django-haystack==2.6.0
djangorestframework==3.6.2
drf-haystack==1.6.1
factory-boy==2.8.1
Faker==0.7.11
geopy==1.11.0
ipaddress==1.0.18
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
Whoosh==2.7.4
